backend/requirements.txt
aiosmtplib==2.0.2
anyio==4.4.0
bcrypt==3.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-mail==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
SQLAlchemy==2.0.30
starlette==0.37.2
typing_extensions==4.12.2
uvicorn[standard]==0.30.1
watchfiles==0.22.0
websockets==12.0
aiosqlite==0.20.0
google-generativeai==0.8.5
greenlet==3.0.3
azure-cognitiveservices-speech==1.38.0
gunicorn
